Wait for connection to Tiller to succeed, on start
In case Tiller is not available at the time the operator starts up,
wait for a `Version` call to succeed.

This also makes the Helm package version explicitly `^2` (anything
that starts with a `2.`); because `dep` implicitly prepends `^` to an
unadorned version, this is the same constraint as before. Though
oddly, Gopkg.lock did not reflect that :-S