Add git.Mirrors, for keeping track of several gits
For the helm operator to be able to sync from git repos mentioned by
charts, rather than from one pre-configured repo, we need to be able
to dynamically start mirroring git repos, and respond when they have
been refreshed.

To that end, this commit introduces git.Mirrors, essentially a map to
which you can idempotently add repos to be mirrored.